move selectors.js to is-requesting/selectors.js
